fix(test): use uint256 instead of uint in SimpleContractFactory
Use explicit uint256 type instead of uint for better code clarity
and consistency with Solidity best practices.

+++ test/datahaven/contracts/src/SimpleContractFactory.sol
@@ -23,7 +23,7 @@ contract SimpleContractFactory {
         deployedWithCreate.push(newContract);
     }
 
-    function createSimpleContractWithCreate2(uint salt) public returns (address) {
+    function createSimpleContractWithCreate2(uint256 salt) public returns (address) {
         bytes memory bytecode = type(SimpleContract).creationCode;
 
         address addr;
@@ -45,4 +45,3 @@ contract SimpleContractFactory {
         return deployedWithCreate2;
     }
 }
-
